MySpringJDBC/src/main/java/com/JDBC/MySpringJDBC/Mapper/ProductMapper.java
package com.JDBC.MySpringJDBC.Mapper;

import com.JDBC.MySpringJDBC.Models.Product;
import org.springframework.jdbc.core.RowMapper;
import org.springframework.stereotype.Component;

import java.sql.ResultSet;
import java.sql.SQLException;

@Component
public class ProductMapper implements RowMapper<Product> {

    @Override
    public Product mapRow(ResultSet rs, int rowNum) throws SQLException {
        return new Product(rs.getInt(1), rs.getString(2), rs.getInt(3), rs.getInt(4), rs.getInt(5));
    }
}
